refactor: separate processor code
